refactor: AgeGroup 도메인 및 repository 패키징 변경
