+++ src/main/java/com/mobble/mobbleserver/infrastructure/persistence/club/ageGroup/JpaAgeGroupRepository.java
@@ -1,15 +1,16 @@
-package com.mobble.mobbleserver.refactor.club.ageGroup.repository;
+package com.mobble.mobbleserver.infrastructure.persistence.club.ageGroup;
 
-import com.mobble.mobbleserver.refactor.club.ageGroup.entity.AgeGroup;
+import com.mobble.mobbleserver.domain.club.ageGroup.AgeGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 
 import java.util.List;
 
-public interface AgeGroupRepository extends JpaRepository<AgeGroup, Long> {
+public interface JpaAgeGroupRepository extends JpaRepository<AgeGroup, Long> {
 
     List<AgeGroup> findByClubId(Long clubId);
 
     @Modifying
     void deleteAllClubAgeGroupByClubId(Long clubId);
 }
+
